refactoring move ScaledLipschitzModule to the modules.module.py to provide the get_scaling_factor and support factory also in LipschitzModule class
